Change package and folder structure
