+++ components/pom.xml
@@ -21,7 +21,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>org.wso2.carbon.identity.conditional.auth.functions</groupId>
+        <groupId>org.wso2.carbon.identity.conditional.fraud.detection.castle</groupId>
         <artifactId>castle-integration</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
@@ -139,11 +139,11 @@
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
                         <Private-Package>
-                            org.wso2.carbon.identity.conditional.auth.functions.internal
+                            org.wso2.carbon.identity.conditional.fraud.detection.castle.internal
                         </Private-Package>
                         <Export-Package>
-                            !org.wso2.carbon.identity.conditional.auth.functions.internal,
-                            org.wso2.carbon.identity.conditional.auth.functions.*;
+                            !org.wso2.carbon.identity.conditional.fraud.detection.castle.internal,
+                            org.wso2.carbon.identity.conditional.fraud.detection.castle.*;
                         </Export-Package>
 
                         <Import-Package>
