+++ components/src/main/java/org/wso2/carbon/identity/conditional/fraud/detection/castle/LoginSuccessRequestSender.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.conditional.auth.functions;
+package org.wso2.carbon.identity.conditional.fraud.detection.castle;
 
 import com.google.common.collect.ImmutableMap;
 import io.castle.client.Castle;
@@ -27,15 +27,16 @@ import io.castle.client.model.CastleRuntimeException;
 import io.castle.client.model.CastleSdkConfigurationException;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.wso2.carbon.identity.conditional.auth.functions.constant.ErrorMessageConstants;
-import org.wso2.carbon.identity.conditional.auth.functions.model.User;
+import org.wso2.carbon.identity.conditional.fraud.detection.castle.constant.ErrorMessageConstants;
+import org.wso2.carbon.identity.conditional.fraud.detection.castle.model.User;
+
 
 /**
- * Class with the Castle request sending functionality for a failed login.
+ * Class with the Castle request sending functionality for a successful login.
  */
-public class LoginFailedRequestSender implements RequestSender {
+public class LoginSuccessRequestSender implements RequestSender {
 
-    private static final Log LOG = LogFactory.getLog(LoginFailedRequestSender.class);
+    private static final Log LOG = LogFactory.getLog(LoginSuccessRequestSender.class);
 
     public CustomCastleResponse doRequest(User user, String castleRequestToken, CastleContext castleContext,
                                           String apiSecret) {
@@ -51,21 +52,21 @@ public class LoginFailedRequestSender implements RequestSender {
         try {
             assert castle != null;
 
-            CastleResponse response = castle.client().filter(ImmutableMap.builder()
-                    .put("type", "$login")
-                    .put("status", "$failed")
-                    .put(Castle.KEY_CONTEXT, ImmutableMap.builder()
-                            .put(Castle.KEY_IP, castleContext.getIp())
-                            .put(Castle.KEY_HEADERS, castleContext.getHeaders())
+            CastleResponse response = castle.client().risk(ImmutableMap.builder()
+                            .put("type", "$login")
+                            .put("status", "$succeeded")
+                            .put(Castle.KEY_CONTEXT, ImmutableMap.builder()
+                                    .put(Castle.KEY_IP, castleContext.getIp())
+                                    .put(Castle.KEY_HEADERS, castleContext.getHeaders())
+                                    .build()
+                            )
+                            .put(Castle.KEY_USER, ImmutableMap.builder()
+                                    .put("id", user.getId())
+                                    .put("email", user.getEmail())
+                                    .build()
+                            )
+                            .put(Castle.KEY_REQUEST_TOKEN, castleRequestToken)
                             .build()
-                    )
-                    .put(Castle.KEY_USER, ImmutableMap.builder()
-                            .put("id", user.getId())
-                            .put("email", user.getEmail())
-                            .build()
-                    )
-                    .put(Castle.KEY_REQUEST_TOKEN, castleRequestToken)
-                    .build()
             );
 
             CustomCastleResponse customCastleResponse = new CustomCastleResponse(response);
