+++ pom.xml
@@ -21,7 +21,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.wso2.carbon.identity.conditional.auth.functions</groupId>
+    <groupId>org.wso2.carbon.identity.conditional.fraud.detection.castle</groupId>
     <artifactId>castle-integration</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>pom</packaging>
